sklep_2/static/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coverage==6.1.1
cryptography==3.4.8
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.7
django-bootstrap-v5==1.0.5
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
fonttools==4.28.0
future==0.18.2
html5lib==1.1
idna==3.2
importlib-metadata==3.10.1
iniconfig==1.1.1
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.31
Pillow==8.3.2
pluggy==1.0.0
polib==1.1.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pydyf==0.1.2
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
pytest==6.2.5
pytest-django==4.4.0
python-bidi==0.4.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.2
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.7
weasyprint==53.3
webencodings==0.5.1
zipp==3.5.0
zopfli==0.1.9
